move tools utils to utils.tools
